Renamed some files for consistency
