--- src/__init__.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
