+++ src/roswire/ros1/node_manager.py
